+++ src/ugdk/base/types.h
@@ -22,22 +22,22 @@ static const Mirror MIRROR_VFLIP  = 2;
 static const Mirror MIRROR_HVFLIP = 3;
 
 typedef struct Color {
-    explicit Color(float _r = 1.0f, float _g = 1.0f, float _b = 1.0f, float _a = 1.0f)
+    explicit Color(double _r = 1.0, double _g = 1.0, double _b = 1.0, double _a = 1.0)
           : r(_r), g(_g), b(_b), a(_a) {}
-    explicit Color(uint32 hex_val, float _a = 1.0f) :
-        r(((hex_val & 0xFF0000) >> 16) / 255.0f),
-        g(((hex_val & 0x00FF00) >>  8) / 255.0f),
-        b(((hex_val & 0x0000FF)      ) / 255.0f),
+    explicit Color(uint32 hex_val, double _a = 1.0) :
+        r(((hex_val & 0xFF0000) >> 16) / 255.0),
+        g(((hex_val & 0x00FF00) >>  8) / 255.0),
+        b(((hex_val & 0x0000FF)      ) / 255.0),
         a(_a) {}
     
 	union {
-        struct { float r, g, b, a; };
-        struct { float val[4];  };
+        struct { double r, g, b, a; };
+        struct { double val[4];  };
     };
 } Color;
 
-//static Color BLACK = {0.0f, 0.0f, 0.0f};
-static const Color WHITE(1.0f, 1.0f, 1.0f);
+//static Color BLACK = {0.0, 0.0, 0.0};
+static const Color WHITE(1.0, 1.0, 1.0);
 
 }  // namespace ugdk
