+++ src/ugdk/graphic/drawable/texturedrectangle.h
@@ -14,7 +14,7 @@ class TexturedRectangle : public Shape {
     TexturedRectangle(Texture* texture, const Vector2D& size);
     ~TexturedRectangle();
 
-    void Draw(float dt);
+    void Draw(double dt);
 
     const Vector2D& size() const { return size_; }
     void set_size(const Vector2D& size) { size_ = size; }
