+++ src/ugdk/math/vector2D.cc
@@ -4,28 +4,28 @@
 namespace ugdk {
 
 // Returns the norm-1.
-float Vector2D::NormOne() const {
+double Vector2D::NormOne() const {
     return fabs(x) + fabs(y);
 }
 
-float Vector2D::Length() const {
+double Vector2D::Length() const {
     return sqrt(LengthSquared());
 }
 
-float Vector2D::Angle() const {
+double Vector2D::Angle() const {
     return atan2(y, x);
 }
 
 Vector2D Vector2D::Normalize() const {
-    float norm = Length();
+    double norm = Length();
     if (norm == 0) {
         return *this;
     }
     return (*this) / norm;
 }
 
-Vector2D Vector2D::Rotate(const float angle) const {
-    float ca = cos(angle), sa = sin(angle);   
+Vector2D Vector2D::Rotate(const double angle) const {
+    double ca = cos(angle), sa = sin(angle);   
     return Vector2D(x * ca - y * sa, x * sa + y * ca);
 }
 
@@ -54,17 +54,17 @@ Vector2D operator-(const Vector2D &left, const Vector2D &right) {
 Vector2D operator-(const Vector2D &left) {
     return Vector2D(-left.x, -left.y);
 }
-Vector2D operator*(const float  &scalar, const Vector2D &right) {
+Vector2D operator*(const double  &scalar, const Vector2D &right) {
     return Vector2D::Multiply(right, scalar);
 }
-Vector2D operator*(const Vector2D &left, const float   &scalar) {
+Vector2D operator*(const Vector2D &left, const double   &scalar) {
     return Vector2D::Multiply(left, scalar);
 }
-Vector2D operator/(const Vector2D &left, const float   &scalar) {
-    return Vector2D::Multiply(left, 1.0f / scalar);
+Vector2D operator/(const Vector2D &left, const double   &scalar) {
+    return Vector2D::Multiply(left, 1.0 / scalar);
 }
 
-float operator*(const Vector2D &left, const Vector2D &right) {
+double operator*(const Vector2D &left, const Vector2D &right) {
     return Vector2D::InnerProduct(left, right);
 }
